+++ src/main/java/roomescape/infrastructure/Jpa/JpaReservationRepository.java
@@ -1,4 +1,4 @@
-package roomescape.infrastructure;
+package roomescape.infrastructure.Jpa;
 
 import org.springframework.context.annotation.Primary;
 import org.springframework.stereotype.Repository;
