+++ src/test/java/roomescape/infrastructure/ReservationRepositoryTest.java
@@ -14,6 +14,10 @@ import roomescape.business.model.repository.ReservationTimeRepository;
 import roomescape.business.model.repository.ThemeRepository;
 import roomescape.business.model.repository.UserRepository;
 import roomescape.business.model.vo.Id;
+import roomescape.infrastructure.Jpa.JpaReservationRepository;
+import roomescape.infrastructure.Jpa.JpaReservationTimeRepository;
+import roomescape.infrastructure.Jpa.JpaThemeRepository;
+import roomescape.infrastructure.Jpa.JpaUserRepository;
 import roomescape.test_util.JpaTestUtil;
 
 import java.time.LocalDate;
