+++ src/test/java/roomescape/infrastructure/ThemeRepositoryTest.java
@@ -8,6 +8,7 @@ import org.springframework.context.annotation.Import;
 import roomescape.business.model.entity.Theme;
 import roomescape.business.model.repository.ThemeRepository;
 import roomescape.business.model.vo.Id;
+import roomescape.infrastructure.Jpa.JpaThemeRepository;
 import roomescape.test_util.JpaTestUtil;
 
 import java.time.LocalDate;
