+++ src/test/java/roomescape/test_util/JpaTestUtil.java
@@ -7,10 +7,10 @@ import roomescape.business.model.entity.ReservationTime;
 import roomescape.business.model.entity.Theme;
 import roomescape.business.model.entity.User;
 import roomescape.business.model.vo.Id;
-import roomescape.infrastructure.JpaReservationDao;
-import roomescape.infrastructure.JpaReservationTimeDao;
-import roomescape.infrastructure.JpaThemeDao;
-import roomescape.infrastructure.JpaUserDao;
+import roomescape.infrastructure.Jpa.JpaReservationDao;
+import roomescape.infrastructure.Jpa.JpaReservationTimeDao;
+import roomescape.infrastructure.Jpa.JpaThemeDao;
+import roomescape.infrastructure.Jpa.JpaUserDao;
 
 import java.time.LocalDate;
 import java.time.LocalTime;
